Move Publisher out into it's own file
